requirements.txt
pygame-ce>=2.5.0
Pillow>=10.0
pytest>=7.0
